Reorganized the modules
